Moved files from render/rendering to render/
Part of ongoing efforts to flatten file structures

The extra nesting is unnecessary since the folder names render and renderer are similar, as well as being confusing as to what they contain.
