+++ include/render/data_renderer.h
@@ -1,7 +1,7 @@
 // This file is subject to the terms and conditions defined in 'LICENSE' in the source code package
 
-#ifndef JACTORIO_INCLUDE_RENDER_RENDERING_DATA_RENDERER_H
-#define JACTORIO_INCLUDE_RENDER_RENDERING_DATA_RENDERER_H
+#ifndef JACTORIO_INCLUDE_RENDERING_DATA_RENDERER_H
+#define JACTORIO_INCLUDE_RENDERING_DATA_RENDERER_H
 #pragma once
 
 #include "core/coordinate_tuple.h"
@@ -26,9 +26,9 @@ namespace jactorio::render
     class RendererLayer;
 
     void DrawConveyorSegmentItems(RendererLayer& layer,
-                                   const SpriteUvCoordsT& uv_coords,
-                                   const core::Position2<float>& pixel_offset,
-                                   game::ConveyorSegment& line_segment);
+                                  const SpriteUvCoordsT& uv_coords,
+                                  const core::Position2<float>& pixel_offset,
+                                  game::ConveyorSegment& line_segment);
 
     void DrawInserterArm(RendererLayer& layer,
                          const SpriteUvCoordsT& uv_coords,
@@ -37,4 +37,4 @@ namespace jactorio::render
                          const proto::InserterData& inserter_data);
 } // namespace jactorio::render
 
-#endif // JACTORIO_INCLUDE_RENDER_RENDERING_DATA_RENDERER_H
+#endif // JACTORIO_INCLUDE_RENDERING_DATA_RENDERER_H
