+++ include/render/gui_renderer.h
@@ -1,7 +1,7 @@
 // This file is subject to the terms and conditions defined in 'LICENSE' included in the source code package
 
-#ifndef JACTORIO_INCLUDE_RENDER_RENDERING_GUI_RENDERER_H
-#define JACTORIO_INCLUDE_RENDER_RENDERING_GUI_RENDERER_H
+#ifndef JACTORIO_INCLUDE_RENDERING_GUI_RENDERER_H
+#define JACTORIO_INCLUDE_RENDERING_GUI_RENDERER_H
 #pragma once
 
 #include <type_traits>
@@ -86,4 +86,4 @@ namespace jactorio::render
 
 } // namespace jactorio::render
 
-#endif // JACTORIO_INCLUDE_RENDER_RENDERING_GUI_RENDERER_H
+#endif // JACTORIO_INCLUDE_RENDERING_GUI_RENDERER_H
