+++ include/render/renderer.h
@@ -1,7 +1,7 @@
 // This file is subject to the terms and conditions defined in 'LICENSE' in the source code package
 
-#ifndef JACTORIO_INCLUDE_RENDER_RENDERING_RENDERER_H
-#define JACTORIO_INCLUDE_RENDER_RENDERING_RENDERER_H
+#ifndef JACTORIO_INCLUDE_RENDERING_RENDERER_H
+#define JACTORIO_INCLUDE_RENDERING_RENDERER_H
 #pragma once
 
 #include <future>
@@ -9,7 +9,7 @@
 
 #include "core/data_type.h"
 #include "render/opengl/mvp_manager.h"
-#include "render/rendering/renderer_layer.h"
+#include "render/renderer_layer.h"
 
 namespace jactorio::game
 {
@@ -202,4 +202,4 @@ namespace jactorio::render
     };
 }; // namespace jactorio::render
 
-#endif // JACTORIO_INCLUDE_RENDER_RENDERING_RENDERER_H
+#endif // JACTORIO_INCLUDE_RENDERING_RENDERER_H
