+++ include/render/renderer_layer.h
@@ -1,7 +1,7 @@
 // This file is subject to the terms and conditions defined in 'LICENSE' in the source code package
 
-#ifndef JACTORIO_INCLUDE_RENDER_RENDERING_RENDERER_LAYER_H
-#define JACTORIO_INCLUDE_RENDER_RENDERING_RENDERER_LAYER_H
+#ifndef JACTORIO_INCLUDE_RENDERING_RENDERER_LAYER_H
+#define JACTORIO_INCLUDE_RENDERING_RENDERER_LAYER_H
 #pragma once
 
 #include <memory>
@@ -180,4 +180,4 @@ namespace jactorio::render
     };
 } // namespace jactorio::render
 
-#endif // JACTORIO_INCLUDE_RENDER_RENDERING_RENDERER_LAYER_H
+#endif // JACTORIO_INCLUDE_RENDERING_RENDERER_LAYER_H
