+++ src/CMakeLists.txt
@@ -92,10 +92,10 @@ set(JACTORIO_SRC_FILES
         ${JACTORIO_DIR}/render/display_window.cpp
         ${JACTORIO_DIR}/render/render_loop.cpp
 
-        ${JACTORIO_DIR}/render/rendering/data_renderer.cpp
-        ${JACTORIO_DIR}/render/rendering/renderer.cpp
-        ${JACTORIO_DIR}/render/rendering/renderer_layer.cpp
-        ${JACTORIO_DIR}/render/rendering/spritemap_generator.cpp
+        ${JACTORIO_DIR}/render/data_renderer.cpp
+        ${JACTORIO_DIR}/render/renderer.cpp
+        ${JACTORIO_DIR}/render/renderer_layer.cpp
+        ${JACTORIO_DIR}/render/spritemap_generator.cpp
         )
 # ======================================== END Source files .cpp
