+++ src/game/player/player_data.cpp
@@ -12,7 +12,7 @@
 #include "proto/recipe.h"
 #include "proto/resource_entity.h"
 #include "proto/tile.h"
-#include "render/rendering/renderer.h"
+#include "render/renderer.h"
 
 using namespace jactorio;
