+++ src/render/display_window.cpp
@@ -17,7 +17,7 @@
 
 #include "gui/imgui_manager.h"
 #include "render/render_loop.h"
-#include "render/rendering/renderer.h"
+#include "render/renderer.h"
 
 using namespace jactorio;
